requirements.txt
# Core ML & Data Science
numpy>=1.22.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.9.0
Pillow>=10.0.0
matplotlib>=3.7.0
pandas>=2.0.0

# YOLO Implementation
ultralytics>=8.1.0

# Image Processing & Metadata
piexif>=1.1.3

# Data Augmentation
albumentations>=2.0.5

# CLI & Utilities
click>=8.1.0
rich>=13.0.0
rich-click>=1.6.0  # Enhanced Click CLI with Rich formatting
pyyaml>=6.0.0
ipython>=8.0.0

# Testing Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
